elt-monorepo-master/shared_libs/requirements.txt
pyodbc
oracledb
pyyaml